refactor: extract formatTime and downloadText into shared utility
