+++ frontend/src/components/ProtocolView/ProtocolCard.tsx
@@ -1,12 +1,5 @@
 import { useStore } from '../../store'
-
-function formatTime(ms: number): string {
-  const s = Math.floor(ms / 1000)
-  const m = Math.floor(s / 60)
-  const h = Math.floor(m / 60)
-  const pad = (n: number) => n.toString().padStart(2, '0')
-  return `${pad(h)}:${pad(m % 60)}:${pad(s % 60)}`
-}
+import { formatTime } from '../../utils/format'
 
 interface Props {
   timestamp: number | null
